Fix slowdown when closing all tabs
We were O^2 stupid: for each tab we were getting its position, which in
turns requires walking the list of children and then we were removing
the tab, which in turns requires gtk to re-walk the list of tabs to find
the one we want to remove. Besides removing tabs starting from the first
is bad for gtknotebook since even/odd css of each tab is recalculated
over and over: we can remove tabs in reverse order.