Notebook menu: fix menu population logic
There were a couple of problems with the logic for menu population.

The biggest problem was that the menu was always being emptied and
repopulated on both open and close, which was needless work.

A more subtle problem, though, was that the menu was being emptied
immediately on the menu being closed.  The "toggled" signal on the
GtkMenuButton was being fired before the action activation had been
delivered.  Removing all items from the menu at this point resulted in
the menu item itself being destroyed, and the signal emission stopping
(so that the action activation never arrived).  Remove the items from
the idle in order to avoid this.

This patch also fixes a leak of GMenuItems when constructing the menu.