ViewFrame: fix search triggering
Trigger the search when the search text is new.

When gtk_text_buffer_select_range() is called, the selection in the
search entry is lost. gtk_text_buffer_select_range() is called when the
search is triggered (i.e. when the buffer is searched).

When the search text is the same, no need to trigger the search, and we
can select the text in the search entry, so the user can search for
something else. On the other hand, when the search text is new, we must
trigger the search, and we are not able to (easily) select the text in
the search entry. But it's not a problem, since the search text is new,
so the user doesn't want to search for something completely different,
he just perhaps wants to adjust a little the search text.