Printing: fix margins
GTK+'s default margins are too small. It was not possible to use a hole
punch, and with some printers the text was truncated on the right.

Add gsettings to configure margins, with better defaults (25mm on each
side). The unit chosen is the millimeter, because it's the international
standard.

https://bugzilla.gnome.org/show_bug.cgi?id=572741