Unset correctly GVFS file attributes
remove_attribute() removes the attribute from the FileInfo, so when
calling g_file_set_attributes_from_info(), the attribute value is kept
as-is, it was not unset.

To unset the file attribute, we must set the attribute to an invalid
type (it seems there is no nicer way to do it).

https://bugzilla.gnome.org/show_bug.cgi?id=741853