Prevent a crash with GeditStatusMenuButton
The problem is not fixed at the source (in GeditWindow). But at least it
will not crash anymore.

GeditWindow dives deep in the containment hierarchy and connects to some
GeditDocument signals. And it seems that a signal is not correctly
disconnected and is called when GeditWindow is disposed. The code in
GeditWindow is messy and there are other problems (getting the buffer
with gtk_text_view_get_buffer() can return another buffer than the one
that the signal was connected to, especially during destroys).

https://bugzilla.gnome.org/show_bug.cgi?id=759174